Updated visualizations
Visualizations are updated, however there may be potential issues with the package-lock file. In the event you run it and there is a missing library, simply install whatever libraries it says may be missing and it should run fine. I will work on getting the package-lock rebuilt and up to date